+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add fabmc_provision_server config to enable/disable server provisioning
+- Add support for provisioning Fabric on Minecraft client
 
 ## 1.2.0 - 2026-02-21
 ### Added